rename NetworkManager to ResourceManager (so we can use content:// urls as well)
