+++ core/src/main/java/com/novoda/pxfetcher/NetworkRetriever.java
@@ -2,7 +2,6 @@ package com.novoda.pxfetcher;
 
 import android.graphics.Bitmap;
 
-import com.novoda.pxfetcher.NetworkManager;
 import com.novoda.pxfetcher.task.Result;
 import com.novoda.pxfetcher.task.Retriever;
 import com.novoda.pxfetcher.task.TagWrapper;
@@ -11,13 +10,13 @@ import java.io.File;
 
 public class NetworkRetriever<T extends TagWrapper<V>, V> implements Retriever<T, V> {
 
-    private final NetworkManager networkManager;
+    private final ResourceManager resourceManager;
     private final FileNameFactory<V> fileNameFactory;
     private final BitmapProcessor bitmapProcessor;
     private final BitmapDecoder decoder;
 
-    public NetworkRetriever(NetworkManager networkManager, FileNameFactory<V> fileNameFactory, BitmapDecoder decoder, BitmapProcessor bitmapProcessor) {
-        this.networkManager = networkManager;
+    public NetworkRetriever(ResourceManager resourceManager, FileNameFactory<V> fileNameFactory, BitmapDecoder decoder, BitmapProcessor bitmapProcessor) {
+        this.resourceManager = resourceManager;
         this.fileNameFactory = fileNameFactory;
         this.bitmapProcessor = bitmapProcessor;
         this.decoder = decoder;
@@ -37,7 +36,7 @@ public class NetworkRetriever<T extends TagWrapper<V>, V> implements Retriever<T
         String sourceUrl = tagWrapper.getSourceUrl();
         String savedUrl = fileNameFactory.getFileName(sourceUrl, tagWrapper.getMetadata());
         File file = new File(savedUrl);
-        networkManager.retrieveImage(sourceUrl, file);
+        resourceManager.retrieveImage(sourceUrl, file);
         return decoder.decode(tagWrapper, file);
     }
