+++ core/src/main/java/com/novoda/pxfetcher/RetrieverFactory.java
@@ -2,8 +2,6 @@ package com.novoda.pxfetcher;
 
 import android.content.res.Resources;
 
-import com.novoda.pxfetcher.CacheManager;
-import com.novoda.pxfetcher.NetworkManager;
 import com.novoda.pxfetcher.task.Retriever;
 import com.novoda.pxfetcher.task.TagWrapper;
 
@@ -12,14 +10,14 @@ public class RetrieverFactory<T extends TagWrapper<V>, V> {
     private static final double DEFAULT_ALLOWED_STRETCHING_THRESHOLD = .10;
     private final CacheManager cacheManager;
     private final FileNameFactory<V> fileManager;
-    private final NetworkManager networkManager;
+    private final ResourceManager resourceManager;
     private final int maxDownsampling;
     private final Resources resources;
 
-    public RetrieverFactory(CacheManager cacheManager, FileNameFactory<V> fileNameFactory, NetworkManager networkManager, int maxDownsampling, Resources resources) {
+    public RetrieverFactory(CacheManager cacheManager, FileNameFactory<V> fileNameFactory, ResourceManager resourceManager, int maxDownsampling, Resources resources) {
         this.cacheManager = cacheManager;
         this.fileManager = fileNameFactory;
-        this.networkManager = networkManager;
+        this.resourceManager = resourceManager;
         this.maxDownsampling = maxDownsampling;
         this.resources = resources;
     }
@@ -39,7 +37,7 @@ public class RetrieverFactory<T extends TagWrapper<V>, V> {
     public Retriever<T, V> createNetworkRetriever() {
         Retriever fileRetriever = createFileRetriever();
         BitmapDecoder croppedDecoder = new DownsamplingBitmapDecoder(resources, DEFAULT_ALLOWED_STRETCHING_THRESHOLD, maxDownsampling);
-        Retriever networkRetriever = new NetworkRetriever<T, V>(networkManager, fileManager, croppedDecoder, new DummyBitmapProcessor());
+        Retriever networkRetriever = new NetworkRetriever<T, V>(resourceManager, fileManager, croppedDecoder, new DummyBitmapProcessor());
         return new FallbackStrategyRetriever<T, V>(fileRetriever, networkRetriever);
     }
